NeuronGrid-Desktop/cluster-core/model-manager/requirements.txt
fastapi
uvicorn
pydantic
requests
huggingface_hub
tqdm
python-dotenv
